cmd/tailscale/tslockjsonv1: extract JSON handling for `tailscale lock`
This patch extracts the JSON handling for the `tailscale lock`
subcommand from the jsonoutput package into its own tslockjsonv1
package.

Updates #17613
