+++ cmd/tailscale/cli/tailnet-lock.go
@@ -24,6 +24,7 @@ import (
 	"github.com/peterbourgon/ff/v3/ffcli"
 
 	"tailscale.com/cmd/tailscale/jsonoutput"
+	"tailscale.com/cmd/tailscale/tslockjsonv1"
 	"tailscale.com/ipn/ipnstate"
 	"tailscale.com/tka"
 	"tailscale.com/tsconst"
@@ -223,7 +224,7 @@ func runTailnetLockStatus(ctx context.Context, args []string) error {
 
 	if nlStatusArgs.json.IsSet {
 		if nlStatusArgs.json.Version == 1 {
-			return jsonoutput.PrintTailnetLockStatusJSONV1(os.Stdout, st)
+			return tslockjsonv1.PrintTailnetLockStatusJSONV1(os.Stdout, st)
 		} else {
 			return fmt.Errorf("unrecognised version: %d", nlStatusArgs.json.Version)
 		}
@@ -716,7 +717,7 @@ func runTailnetLockLog(ctx context.Context, args []string) error {
 func printTailnetLockLog(updates []ipnstate.NetworkLockUpdate, out io.Writer, jsonSchema jsonoutput.SchemaVersion, useColor bool) error {
 	if jsonSchema.IsSet {
 		if jsonSchema.Version == 1 {
-			return jsonoutput.PrintTailnetLockLogJSONV1(out, updates)
+			return tslockjsonv1.PrintTailnetLockLogJSONV1(out, updates)
 		} else {
 			return fmt.Errorf("unrecognised version: %d", jsonSchema.Version)
 		}
