+++ cmd/tailscale/cli/tailnet-lock_test.go
@@ -12,6 +12,7 @@ import (
 	"go4.org/mem"
 
 	"tailscale.com/cmd/tailscale/jsonoutput"
+	"tailscale.com/cmd/tailscale/tslockjsonv1"
 	"tailscale.com/ipn/ipnstate"
 	"tailscale.com/tailcfg"
 	"tailscale.com/tka"
@@ -356,7 +357,7 @@ func TestTailnetLockStatusOutput(t *testing.T) {
 				t.Parallel()
 
 				var outBuf bytes.Buffer
-				err := jsonoutput.PrintTailnetLockStatusJSONV1(&outBuf, &tt.Status)
+				err := tslockjsonv1.PrintTailnetLockStatusJSONV1(&outBuf, &tt.Status)
 				if err != nil {
 					t.Fatalf("PrintTailnetLockStatusJSONV1: %v", err)
 				}
