+++ cmd/tailscale/depaware.txt
@@ -186,8 +186,9 @@ tailscale.com/cmd/tailscale dependencies: (generated by github.com/tailscale/dep
         tailscale.com/cmd/tailscale/cli                              from tailscale.com/cmd/tailscale
         tailscale.com/cmd/tailscale/cli/ffcomplete                   from tailscale.com/cmd/tailscale/cli
         tailscale.com/cmd/tailscale/cli/ffcomplete/internal          from tailscale.com/cmd/tailscale/cli/ffcomplete
-        tailscale.com/cmd/tailscale/jsonoutput                       from tailscale.com/cmd/tailscale/cli
+        tailscale.com/cmd/tailscale/jsonoutput                       from tailscale.com/cmd/tailscale/cli+
         tailscale.com/cmd/tailscale/tsdnsjsonv0                      from tailscale.com/cmd/tailscale/cli
+        tailscale.com/cmd/tailscale/tslockjsonv1                     from tailscale.com/cmd/tailscale/cli
         tailscale.com/cmd/tailscale/tsroutecheckjsonv0               from tailscale.com/cmd/tailscale/cli
         tailscale.com/control/controlbase                            from tailscale.com/control/controlhttp+
         tailscale.com/control/controlhttp                            from tailscale.com/control/ts2021
