+++ cmd/tailscale/tslockjsonv1/doc.go
@@ -1,0 +1,8 @@
+// Copyright (c) Tailscale Inc & contributors
+// SPDX-License-Identifier: BSD-3-Clause
+
+//go:build !ts_omit_tailnetlock
+
+// Package tslockjsonv1 provides types for unmarshaling the JSON output of the
+// "tailscale lock --json=1" command.
+package tslockjsonv1
