+++ cmd/tailscaled/depaware.txt
@@ -267,7 +267,8 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
         tailscale.com/client/web                                     from tailscale.com/ipn/ipnlocal
         tailscale.com/clientupdate                                   from tailscale.com/feature/clientupdate
   LW    tailscale.com/clientupdate/distsign                          from tailscale.com/clientupdate
-        tailscale.com/cmd/tailscale/jsonoutput                       from tailscale.com/feature/tailnetlock
+        tailscale.com/cmd/tailscale/jsonoutput                       from tailscale.com/cmd/tailscale/tslockjsonv1
+        tailscale.com/cmd/tailscale/tslockjsonv1                     from tailscale.com/feature/tailnetlock
         tailscale.com/cmd/tailscaled/childproc                       from tailscale.com/cmd/tailscaled+
         tailscale.com/cmd/tailscaled/tailscaledhooks                 from tailscale.com/cmd/tailscaled+
         tailscale.com/control/controlbase                            from tailscale.com/control/controlhttp+
